feat(overlays): add onMounted hook to hide overflow in chat
This commit includes changes in 'chat.vue' component to hide the body overflow. It does this by adding the 'onMounted' method from vue's lifecycle hooks which sets the body's overflow style to hidden once the component mounts. This provides better UX by preventing the accidental scrolling of the main body when the chat overlay is active.