fix: add condition to prevent triggering kappagen for its own command
Kappagen event was previously being triggered event in response to its own command. This has been corrected with this commit. A condition has been added within apps/events/internal/grpc_impl/impl.go to check if the default command is "kappagen". If it isn't, the kappagen event will be triggered, but if it is, the event will be bypassed.