Refine Dockerfiles and include Go Modules
This commit refines the Dockerfiles for several apps including api, bots, discord, emotes cacher, events, eventsub, parser, scheduler, timers, tokens, websockets, ytsr, and a base Dockerfile. The go mod download command was removed from the individual app Dockerfiles and relocated to the base Dockerfile to optimize the building process. All the go mod files in the apps and libraries directories are now copied to the base Dockerfile for the go mod download to be performed once for all apps. Additionally, the gen-dockerfile-copy tool was updated to reflect these changes and help generate copy commands for the Dockerfile.