chore: new docker layout with self-hosted runners (#444)
* new docker layout

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* fix(websockets): add missed `RefreshKappagenOverlaySettings`

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* Refine Dockerfiles and include Go Modules

This commit refines the Dockerfiles for several apps including api, bots, discord, emotes cacher, events, eventsub, parser, scheduler, timers, tokens, websockets, ytsr, and a base Dockerfile. The go mod download command was removed from the individual app Dockerfiles and relocated to the base Dockerfile to optimize the building process. All the go mod files in the apps and libraries directories are now copied to the base Dockerfile for the go mod download to be performed once for all apps. Additionally, the gen-dockerfile-copy tool was updated to reflect these changes and help generate copy commands for the Dockerfile.

* "Refactor Dockerfile copy commands and cleanup workflow

The Dockerfile copy commands were refactored to improve clarity and reduce redundancy. The copied files no longer deposit in the root directory. Generated copy commands are now enclosed between START COPYGEN and END COPYGEN comments to ensure maintainability.

As part of the same commit, a cleanup step was added to the GitHub Actions workflow. It leverages TooMuch4U's runner-post-cleanup after the cache and build steps to ensure a clean working environment for subsequent runs."

* "Refactor Dockerfile copy commands and cleanup workflow

The Dockerfile copy commands were refactored to improve clarity and reduce redundancy. The copied files no longer deposit in the root directory. Generated copy commands are now enclosed between START COPYGEN and END COPYGEN comments to ensure maintainability.

As part of the same commit, a cleanup step was added to the GitHub Actions workflow. It leverages TooMuch4U's runner-post-cleanup after the cache and build steps to ensure a clean working environment for subsequent runs."

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto

* kruto